Extract pattern matching functions and tests to matcher.go
